Patches for Safari (#46)
I tested the site on my older brother's MacBook.

And woah, Safari is a different browser. Like it's wild. The dev tools
have so many missing features, and a few design decision that make me
question my sanity.

Anyways, this PR does
- Bumps the version and the dependencies.
- Replaces some emojis, and removes the Noto Color Emojis font. Makes
the site smoler, and fixes the major slowness on Safari. I tested it on
Windows 10, which sometimes slightly sucks in terms of emoji support.
- Adds two lines to the CSS to tell Safari to get its stuff together and
render the tiles properly